+++ src/app/index.tsx
@@ -1,7 +1,6 @@
 import { RouterProvider } from 'react-router-dom';
 
 import { AuthProvider, ErrorBoundary, ThemeProvider } from './providers';
-import { THEMES } from './providers/theme/locales';
 import { router } from './router';
 
 import { Toaster } from '@/shared/ui/sonner';
@@ -9,7 +8,7 @@ import { Toaster } from '@/shared/ui/sonner';
 export const App = () => {
   return (
     <ErrorBoundary>
-      <ThemeProvider defaultTheme={THEMES.DARK} storageKey="project-theme">
+      <ThemeProvider>
         <AuthProvider>
           <RouterProvider router={router} />
         </AuthProvider>
