+++ src/app/providers/index.ts
@@ -1,5 +1,3 @@
 export { AuthProvider } from './auth-provider';
 export { ErrorBoundary } from './error-boundary';
-export { ThemeProviderContext } from './theme/theme-context';
 export { ThemeProvider } from './theme/theme-provider';
-export * from './theme/locales';
