--- src/app/providers/theme/locales/index.ts
@@ -1,9 +1,0 @@
-export const THEMES = {
-  DARK: 'dark',
-  LIGHT: 'light',
-  SYSTEM: 'system',
-};
-
-export const THEMES_ERRORS = {
-  NOT_WITHIN_THEME_PROVIDER: 'useTheme должен использоваться внутри ThemeProvider',
-};
