+++ src/app/ui/header/locales/index.ts
@@ -5,5 +5,4 @@ export const MENU_LINK_TEXT = {
   PROFILE: 'Профиль',
   LOGOUT: 'Выход',
   APP_TITLE: 'Тренажёр интервью',
-  THEME: 'Тема оформления',
 };
