+++ src/app/ui/header/ui/menu-links.tsx
@@ -5,7 +5,7 @@ import { MENU_LINK_TEXT } from '../locales';
 
 import { MENU_LINKS } from './lib/constants';
 
-import { ModeToggle } from '@/features/theme/ui/mode-toggle';
+import { ModeToggle } from '@/features/theme';
 import { Button } from '@/shared';
 import { signOut } from '@/shared/api';
 import { cn } from '@/shared/lib/utils';
@@ -81,7 +81,7 @@ export const MenuLinks = ({ variant, mobileOpen, onNavigate }: MenuLinksProps) =
           isDesktop ? 'ml-2' : 'px-4 py-2 border-t border-white/10 mt-1 justify-between'
         )}
       >
-        {!isDesktop && <span className="text-sm text-light/60">MENU_LINK_TEXT.THEME</span>}
+        {!isDesktop && <span className="text-sm text-light/60"></span>}
         <ModeToggle />
       </div>
