+++ src/features/theme/index.ts
@@ -1,0 +1,1 @@
+export { ModeToggle } from './ui/mode-toggle';
