+++ src/features/theme/ui/mode-toggle.tsx
@@ -1,41 +1,26 @@
 import { Moon, Sun } from 'lucide-react';
 
-import { THEMES } from '@/app/providers';
-import { Button, useTheme } from '@/shared';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@/shared/ui/dropdown-menu';
+import { useTheme } from '@/shared/hooks/use-theme';
+import { THEMES } from '@/shared/lib/constants';
+import { Button } from '@/shared/ui/button';
 
 export const ModeToggle = () => {
-  const { setTheme } = useTheme();
+  const { theme, setTheme } = useTheme();
+
+  const toggleTheme = () => {
+    setTheme(theme === THEMES.DARK ? THEMES.LIGHT : THEMES.DARK);
+  };
 
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="ghost"
-          size="icon"
-          className="text-light hover:bg-white/10 focus-visible:ring-0"
-        >
-          <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
-          <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
-          <span className="sr-only">Toggle theme</span>
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="bg-card border-white/10 text-light">
-        <DropdownMenuItem className="focus:bg-white/10" onClick={() => setTheme(THEMES.LIGHT)}>
-          Light
-        </DropdownMenuItem>
-        <DropdownMenuItem className="focus:bg-white/10" onClick={() => setTheme(THEMES.DARK)}>
-          Dark
-        </DropdownMenuItem>
-        <DropdownMenuItem className="focus:bg-white/10" onClick={() => setTheme(THEMES.SYSTEM)}>
-          System
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
+    <Button
+      variant="ghost"
+      size="icon"
+      onClick={toggleTheme}
+      className="text-light hover:bg-white/10 focus-visible:ring-0"
+    >
+      <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
+      <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
+      <span className="sr-only">Toggle theme</span>
+    </Button>
   );
 };
