+++ src/shared/hooks/use-theme.ts
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
 
-import { THEMES_ERRORS, ThemeProviderContext } from '@/app/providers/';
+import { NOT_WITHIN_THEME_PROVIDER_ERROR, ThemeProviderContext } from '@/shared';
 
 export const useTheme = () => {
   const context = useContext(ThemeProviderContext);
 
-  if (context === undefined) throw new Error(THEMES_ERRORS.NOT_WITHIN_THEME_PROVIDER);
+  if (context === undefined) throw new Error(NOT_WITHIN_THEME_PROVIDER_ERROR);
 
   return context;
 };
