+++ src/shared/index.ts
@@ -10,6 +10,9 @@ export {
   STAGES,
   DEFAULT_STAGES_VALUE,
   PASSING_SCORE,
+  THEMES,
+  THEME_STORAGE_KEY,
+  type Theme,
 } from './lib/constants';
 
 export { getProgressPercent } from './lib/get-progress-percent';
@@ -17,3 +20,5 @@ export { isTopicCompleted } from './lib/is-topic-completed';
 export { formatScore } from './lib/format-score';
 export { groupByStage } from './lib/group-by-stage';
 export { getTopicStats } from './lib/get-topic-stats';
+export { NOT_WITHIN_THEME_PROVIDER_ERROR } from './lib/locales';
+export { ThemeProviderContext } from './model/theme/theme-context';
