+++ src/shared/lib/constants.ts
@@ -23,3 +23,13 @@ export const STAGES = [
 
 export const DEFAULT_STAGES_VALUE = STAGES[0].id;
 export const PASSING_SCORE = 70;
+
+export const THEMES = {
+  DARK: 'dark',
+  LIGHT: 'light',
+  SYSTEM: 'system',
+} as const;
+
+export type Theme = (typeof THEMES)[keyof typeof THEMES];
+
+export const THEME_STORAGE_KEY = 'project-theme';
