+++ src/shared/lib/locales.ts
@@ -4,3 +4,5 @@ export const STAGE_3_TITLE = 'Этап 3';
 export const STAGE_1_DESCRIPTION = 'Технический скрининг';
 export const STAGE_2_DESCRIPTION = 'Основы JS 1';
 export const STAGE_3_DESCRIPTION = 'Основы JS 2';
+export const NOT_WITHIN_THEME_PROVIDER_ERROR =
+  'useTheme должен использоваться внутри ThemeProvider';
